requirements.txt
python-yaml
python-requests
python-lxml
python-cssselect
